Blocks refactor
