+++ cms/navigation/blocks.py
@@ -1,0 +1,48 @@
+from wagtail.blocks import CharBlock, ListBlock, PageChooserBlock, StructBlock
+from cms.core.blocks.base import LinkBlock
+from django.utils.translation import gettext_lazy as _
+
+
+class ThemeLinkBlock(LinkBlock):
+    page = PageChooserBlock(required=False, page_type="themes.ThemePage")
+
+    class Meta:
+        label = _("Theme Link")
+
+
+class TopicLinkBlock(LinkBlock):
+    page = PageChooserBlock(required=False, page_type="topics.TopicPage")
+
+    class Meta:
+        label = _("Topic Link")
+
+
+class HighlightsBlock(LinkBlock):
+    description = CharBlock(
+        required=True, max_length=50, help_text=_("For example: 'View our latest and upcoming releases.'")
+    )
+
+    class Meta:
+        icon = "star"
+        label = _("Highlight")
+
+
+class SectionBlock(StructBlock):
+    section_link = ThemeLinkBlock(help_text=_("Main link for this section (Theme pages or external URLs)."))
+    links = ListBlock(
+        TopicLinkBlock(),
+        help_text=_("Sub-links for this section (Topic pages or external URLs)."),
+        max_num=15,
+    )
+
+    class Meta:
+        icon = "folder"
+        label = _("Section")
+
+
+class ColumnBlock(StructBlock):
+    sections = ListBlock(SectionBlock(), label="Sections", max_num=3)
+
+    class Meta:
+        icon = "list-ul"
+        label = _("Column")
